Added dedicated threadpool supervisor.
The supervisor polls the workers every x seconds to see if they have updated their status flag.
The status flag change is scheduled to each worker as an action.
When the action run, the status is updated, if the worker is frozen, the action will never run and the status will never be updated.

If the worker status is unchanged, the supervisor will abort the worker thread and request a new one.

TODO: add tests for this
TODO: fix existing "the world should not end" test